+++ CHANGES.md
@@ -10,6 +10,7 @@
 - macOS: Point the user to the OS' please-enable-notification prompt #474 
 - Improve error alerts #489
 - Update WireGuardKit to fix crash in tunnel
+- Update logger to use file handles to fix crash in tunnel
 
 ## 3.0.6